+++ packages/at_secondary_server/lib/src/enroll/enroll_datastore_value.g.dart
@@ -26,6 +26,7 @@ EnrollDataStoreValue _$EnrollDataStoreValueFromJson(
           Duration(milliseconds: json['apkamKeysExpiryInMillis'] ?? 0)
       ..metadata = (json['metadata'] as Map<String, dynamic>?)
       ..signingAlgo = json['signingAlgo'] as String?
+      ..apsk = (json['apsk'] as Map<String, dynamic>?)
       ..parentEnrollmentId = json['parentEnrollmentId'] as String?;
 
 Map<String, dynamic> _$EnrollDataStoreValueToJson(
@@ -43,6 +44,7 @@ Map<String, dynamic> _$EnrollDataStoreValueToJson(
           instance.apkamKeysExpiryDuration.inMilliseconds,
       if (instance.metadata != null) 'metadata': instance.metadata,
       if (instance.signingAlgo != null) 'signingAlgo': instance.signingAlgo,
+      if (instance.apsk != null) 'apsk': instance.apsk,
       if (instance.parentEnrollmentId != null)
         'parentEnrollmentId': instance.parentEnrollmentId,
     };